fixing the duplicate code issue
